chaos: move comprehensive test under test directory
