+++ test/create-root-graph.spec.ts
@@ -1,7 +1,5 @@
-import { attribute } from './attribute.js';
-import { Graph, Digraph } from './models/index.js';
-import { digraph, graph, strict } from './create-root-graph.js';
-import { toDot } from './to-dot.js';
+import { digraph, graph, strict, Graph, Digraph, attribute as _ } from 'ts-graphviz';
+import { toDot } from '#test/utils';
 
 describe('function digraph', () => {
   it('should return Graph object, when execute digraph()', () => {
@@ -21,7 +19,7 @@ describe('function digraph', () => {
           id: 'id',
           strictMode: false,
           g: digraph('id', {
-            [attribute.label]: 'Label',
+            [_.label]: 'Label',
           }),
         },
       ],
@@ -50,7 +48,7 @@ describe('function digraph', () => {
           id: undefined,
           strictMode: false,
           g: digraph({
-            [attribute.label]: 'Label',
+            [_.label]: 'Label',
           }),
         },
       ],
@@ -61,7 +59,7 @@ describe('function digraph', () => {
           id: undefined,
           strictMode: true,
           g: strict.digraph({
-            [attribute.label]: 'Label',
+            [_.label]: 'Label',
           }),
         },
       ],
@@ -72,7 +70,7 @@ describe('function digraph', () => {
           id: 'id',
           strictMode: true,
           g: strict.digraph('id', {
-            [attribute.label]: 'Label',
+            [_.label]: 'Label',
           }),
         },
       ],
@@ -83,7 +81,7 @@ describe('function digraph', () => {
           id: undefined,
           strictMode: true,
           g: strict.digraph({
-            [attribute.label]: 'Label',
+            [_.label]: 'Label',
           }),
         },
       ],
@@ -126,19 +124,19 @@ describe('function digraph', () => {
       const b = g.node('bb');
       const c = g.node('cc');
       g.edge([a, b, c], {
-        [attribute.color]: 'red',
+        [_.color]: 'red',
       });
       g.subgraph('A', (A) => {
         const Aa = A.node('Aaa', {
-          [attribute.color]: 'pink',
+          [_.color]: 'pink',
         });
 
         const Ab = A.node('Abb', {
-          [attribute.color]: 'violet',
+          [_.color]: 'violet',
         });
         const Ac = A.node('Acc');
         A.edge([Aa.port('a'), Ab, Ac, 'E'], {
-          [attribute.color]: 'red',
+          [_.color]: 'red',
         });
       });
     });
@@ -235,19 +233,19 @@ describe('function graph', () => {
       const b = g.node('bb');
       const c = g.node('cc');
       g.edge([a, b, c], {
-        [attribute.color]: 'red',
+        [_.color]: 'red',
       });
       g.subgraph('A', (A) => {
         const Aa = A.node('Aaa', {
-          [attribute.color]: 'pink',
+          [_.color]: 'pink',
         });
 
         const Ab = A.node('Abb', {
-          [attribute.color]: 'violet',
+          [_.color]: 'violet',
         });
         const Ac = A.node('Acc');
         A.edge([Aa.port('a'), Ab, Ac, 'E'], {
-          [attribute.color]: 'red',
+          [_.color]: 'red',
         });
       });
     });
