+++ tsconfig.json
@@ -12,8 +12,10 @@
     "declaration": true,
     "forceConsistentCasingInFileNames": true,
     "paths": {
-      "#lib/*": ["./src/*/index.js"]
+      "#lib/*": ["./src/*/index.js"],
+      "#test/*": ["./test/*/index.js"],
+      "ts-graphviz": ["./src/index.js"],
+      "ts-graphviz/ast": ["./src/ast/index.js"]
     }
-  },
-  "include": ["src/**/*.ts"]
+  }
 }
